تغییرات ظاهری

+++ ViewModels/Payment/InvestmentSummaryViewModel.cs
@@ -14,7 +14,7 @@ namespace ViewModels
         public int ActiveUsers { get; set; }
         public int InvestmentSuccessPlanCount { get; set; }
         public int? InvestmentCountPerson { get; set; }
-        public long TotalDepositToInvestors { get; set; }
+        public double TotalDepositToInvestors { get; set; }
         public int CountDepositToInvestors { get; set; }
